requirements.txt
appdirs==1.4.4
asgiref==3.4.1
biopython==1.78
certifi==2020.12.5
chardet==4.0.0
crispy-bootstrap5==0.4
decorator==4.4.2
Django==3.2.5
django-crispy-forms==1.12.0
entrezpy==2.1.1
idna==2.10
networkx==2.5
numpy==1.19.5
prettytable==2.0.0
primer3-py==0.6.1
pydna==3.1.3
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
wcwidth==0.2.5
